Create Merge on Nested Column
The task: Update a business API and transferring data. Old data stored in CSV with limited columns. New data stored in JSON with matching record values in a nested column. I extracted the ID string value using a loop function, and created a new column for those values. Finally, I merged the columns and wrote the new data to a CSV.